자신의 게시물만 삭제 가능

+++ backend/src/main/java/com/example/backend/board/service/BoardService.java
@@ -1,7 +1,7 @@
 package com.example.backend.board.service;
 
 import com.example.backend.board.dto.BoardDto;
-import com.example.backend.board.dto.BoardListInfo;
+import com.example.backend.board.dto.BoardListDto;
 import com.example.backend.board.entity.Board;
 import com.example.backend.board.repository.BoardRepository;
 import com.example.backend.member.entity.Member;
@@ -49,24 +49,34 @@ public class BoardService {
         return true;
     }
 
-    public List<BoardListInfo> list() {
-        return boardRepository.findAllByOrderByIdDesc();
+    public List<BoardListDto> list() {
+//        return boardRepository.findAllByOrderByIdDesc();
+        return boardRepository.findAllBy();
     }
 
 
     public BoardDto getBoardById(Integer id) {
-        Board board = boardRepository.findById(id).get();
-        BoardDto boardDto = new BoardDto();
-        boardDto.setId(board.getId());
-        boardDto.setTitle(board.getTitle());
-        boardDto.setContent(board.getContent());
-        boardDto.setAuthor(board.getAuthor());
-        boardDto.setInsertedAt(board.getInsertedAt());
+        BoardDto board = boardRepository.findBoardById(id);
+//        BoardDto boardDto = new BoardDto();
+//        boardDto.setId(board.getId());
+//        boardDto.setTitle(board.getTitle());
+//        boardDto.setContent(board.getContent());
+//        boardDto.setAuthor(board.getAuthor());
+//        boardDto.setInsertedAt(board.getInsertedAt());
+
         return boardDto;
     }
 
-    public void deleteById(Integer id) {
-        boardRepository.deleteById(id);
+    public void deleteById(Integer id, Authentication authentication) {
+        if (authentication == null) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+        Board db = boardRepository.findById(id).get();
+        if (db.getAuthor().getEmail().equals(authentication.getName())) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("권한이 없습니다.");
+        }
     }
 
     public void update(BoardDto boardDto) {
